Add connection mechanism for ibapir
Add ib_connect function to setup the connection to the ib tws or gateway.
Add functions to disconnect and to do some checks if connection is open (is_ib_connected) and a check to see if it is a valid connection (is_ib_connection).

Add reqIds function and the processNextValidIdMsg function for processing the nextvalidid message

Add reqManagedAccts and processManagedAcctsMsg for processing the managed account messages.

Add the ProcessMsg function to decode the incoming messages. This function swtiches between processing functions based on the incoming message id. An escape is built in to print any unknown messages to the console.

Add the processErrMsg to be able to process any error messages. At the moment this only prints the messages.

Add documentation for ib_connect, reqIds and reqManagedAccts (and ProcessMsg but this is set to internal).

Update to namespace to include the new exported functions.