Add 2 new tests to create_placeOrder_msg
Add test for message length. This will need to be expanded to check for orders that add legs to the message

Add test for version check. From server version 145 the version field is no longer sent. Build test to check if a new (higher) server version is released the test still passes. If not, either a new field is being introduced or the version field is being used again.